+++ sdk/containerservice/arm-containerservice/src/api/machines/options.ts
@@ -7,4 +7,14 @@ import type { OperationOptions } from "@azure-rest/core-client";
 export interface MachinesListOptionalParams extends OperationOptions {}
 
 /** Optional parameters. */
+export interface MachinesCreateOrUpdateOptionalParams extends OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** The request should only proceed if an entity matches this string. */
+  ifMatch?: string;
+  /** The request should only proceed if no entity matches this string. */
+  ifNoneMatch?: string;
+}
+
+/** Optional parameters. */
 export interface MachinesGetOptionalParams extends OperationOptions {}